+++ CHANGELOG.md
@@ -12,10 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas_occ.geometry.OCCRevolutionSurface`.
 * Added `compas_occ.conversions.compas_axis_to_occ_axis`.
 * Added `compas_occ.conversions.compas_axis_from_occ_axis`.
-* Added `compas_occ.geometry.OCCCurve.to_polyline`.
 * Added `compas_occ.geometry.OCCExtrusionSurface`.
 * Added `compas_occ.geometry.OCCNurbsSurface.from_extrusion`.
-* Added `compas_occ.geometry.OCCCurve.divide`.
+* Added implementation for `compas_occ.geometry.OCCCurve.divide_by_count`.
+* Added `compas_occ.geometry.OCCCurve.divide` as alias for `compas_occ.geometry.OCCCurve.divide_by_count`.
 
 ### Changed